+++ pkgs/test_core/CHANGELOG.md
@@ -3,6 +3,7 @@
 * Remove support for `FORCE_TEST_EXIT`.
 * Report incomplete tests as errors in the JSON reporter when the run is
   canceled early.
+* Don't log the --test-randomization-ordering-seed if using the json reporter.
 
 ## 0.3.29